+++ Cargo.toml
@@ -15,4 +15,3 @@ salsa = "*"
 ariadne = "*"
 yansi = "*"
 rowan = "*"
-atomic_once_cell = "*"
