+++ tests/DepErrors.pk
@@ -1,0 +1,12 @@
+type Data of
+    New I32
+
+fun borrow(x: &Data): () = ()
+fun move(x: Data): () = ()
+
+fun fail1(x: Data): &Data = do
+    let y: &Data = x # x is borrowed here
+    let z = y
+    let w = z # w depends on x through z
+    move(x) # x is moved here
+    w # error: w depends on x whichH has been moved
